Change locked collateral from Int to Decimal

+++ fathomapp-subgraph/src/position-manager.ts
@@ -27,7 +27,7 @@ export function newPositionHandler(event: LogNewPosition): void {
     position.walletAddress = event.params._own;
     position.collateralPool = poolId
     position.collateralPoolName = poolId.toString()
-    position.lockedCollateral = BigInt.fromString('0')
+    position.lockedCollateral = BigDecimal.fromString('0')
     position.debtShare = BigInt.fromString('0')
     position.safetyBuffer = BigDecimal.fromString('1')
     position.safetyBufferInPercent = BigDecimal.fromString('0')
